+++ src/Resources/contao/languages/de/default.php
@@ -1,3 +1,91 @@
 <?php
 
 $GLOBALS['TL_LANG']['Fideid']['emailCorrupt'] = 'Die Adresse %s enthält ungültige Zeichen!';
+
+// Antragsformular
+$GLOBALS['TL_LANG']['fideid-form']['submit'] = array('Weiter', 'Weiter');
+
+// Seite 1
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite1'] = '<h2>Antragsformular Seite 1</h2><p>Mit dem Abschicken dieser Formularseite wird ein Datensatz auf unserem Server angelegt!</p>';
+$GLOBALS['TL_LANG']['fideid-form']['germany'] = array('Länderkennung GER', 'Soll die Person, für die eine FIDE-Identifikationsnummer beantragt wird, unter Deutschland bei der FIDE geführt werden? Falls nicht, kann der Antrag nicht weiter bearbeitet werden. Bitte wenden Sie sich an die zuständige Föderation. Die Adressen finden Sie im <a href="https://www.fide.com/directory/member-federations" target="_blank">FIDE-Verzeichnis</a>.');
+$GLOBALS['TL_LANG']['fideid-form']['datenschutz'] = array('Datenweitergabe an die FIDE', 'Ich willige ein, dass der Deutsche Schachbund die Daten für die beantragte Person an die FIDE weitergibt.');
+$GLOBALS['TL_LANG']['fideid-form']['ueber18'] = array('Ü18-Altersabfrage', 'Wenn Sie das nachfolgende Formular ausfüllen möchten, müssen Sie über 18 Jahre alt sein. Bitte bestätigen Sie uns, das Sie über 18 Jahre alt sind.');
+
+// Seite 2
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite2'] = '<h2>Antragsformular Seite 2</h2><p>Für wen soll die FIDE-Identifikationsnummer beantragt werden?</p>';
+$GLOBALS['TL_LANG']['fideid-form']['nachname'] = array('Nachname', 'Nachname');
+$GLOBALS['TL_LANG']['fideid-form']['vorname'] = array('Vorname', 'Vorname');
+$GLOBALS['TL_LANG']['fideid-form']['email'] = array('E-Mail', 'E-Mail');
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite2_email'] = '<p>Tragen Sie bitte Ihre E-Mail-Adresse ein, wenn Sie die FIDE-Identifikationsnummer für einen Minderjährigen beantragen.</p>';
+$GLOBALS['TL_LANG']['fideid-form']['titel'] = array('Titel', 'Akademischer Titel');
+$GLOBALS['TL_LANG']['fideid-form']['titel_optionen'] = array('Dr.' => 'Dr.', 'Prof.' => 'Prof.', 'Prof.Dr.' => 'Prof.Dr.');
+$GLOBALS['TL_LANG']['fideid-form']['geburtsdatum'] = array('Geburtsdatum', 'Geburtsdatum');
+$GLOBALS['TL_LANG']['fideid-form']['geschlecht'] = array('Geschlecht', 'Geschlecht');
+$GLOBALS['TL_LANG']['fideid-form']['geschlecht_optionen'] = array('M' => 'Männlich', 'W' => 'Weiblich');
+
+// Seite 3
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite3'] = '<h2>Antragsformular Seite 2.1</h2><p>Die beantragte Person ist minderjährig. Wir benötigen jetzt Daten von Ihnen, als erziehungsberechtigte Person.</p>';
+$GLOBALS['TL_LANG']['fideid-form']['nachname_person'] = array('Nachname', 'Nachname');
+$GLOBALS['TL_LANG']['fideid-form']['vorname_person'] = array('Vorname', 'Vorname');
+$GLOBALS['TL_LANG']['fideid-form']['email_person'] = array('E-Mail', 'E-Mail');
+$GLOBALS['TL_LANG']['fideid-form']['ausweis'] = array('Personalausweis oder Geburtsurkunde', '');
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite3_ausweis'] = '<p>Die beantragte Person ist unter 18 Jahre alt. Der Deutsche Schachbund benötigt eine Kopie des Personalausweises, falls vorhanden. Andernfalls laden Sie eine Kopie der Geburtsurkunde hier hoch. Erlaubt sind die Formate JPG, JPEG, GIF, PNG oder PDF.</p>';
+
+// Seite 4
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite4'] = '<h2>Antragsformular Seite 3</h2>';
+$GLOBALS['TL_LANG']['fideid-form']['turnier'] = array('Turnier', 'Turnier');
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite4_turnier'] = '<p>Für welches Turnier oder welche Veranstaltung wird die FIDE-Identifikationsnummer benötigt?<br>Statt eines Turniers können Sie hier auch einen anderen Grund eingeben, z.B. <i>FIDE-Lizenz Schiedsrichter</i> oder <i>Organisator Elo-Turnier</i>.</p>';
+$GLOBALS['TL_LANG']['fideid-form']['imVerein'] = array('Mitglied im Schachverein', 'Die beantragte Person ist Mitglied in einem Schachverein des Deutschen Schachbundes.');
+
+// Seite 5
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite5'] = '<h2>Antragsformular Seite 3.1</h2>';
+$GLOBALS['TL_LANG']['fideid-form']['verein'] = array('Vereinsname', 'Vereinsname');
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite5_verein'] = '<p>Wie heißt der DSB-Schachverein, in dem die beantragte Person Mitglied ist?</p>';
+
+// Seite 6
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite6'] = '<h2>Antragsformular Seite 3.1</h2>';
+$GLOBALS['TL_LANG']['fideid-form']['ausweis'] = array('Personalausweis oder Geburtsurkunde', '');
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite6_ausweis'] = '<p>Die beantragte Person ist kein Vereinsmitglied. Der Deutsche Schachbund benötigt eine Kopie des Personalausweises, falls vorhanden. Andernfalls laden Sie eine Kopie der Geburtsurkunde hier hoch. Erlaubt sind die Formate JPG, JPEG, GIF, PNG oder PDF.</p>';
+
+// Seite 7
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite7'] = '<h2>Antragsformular Seite 4</h2>';
+$GLOBALS['TL_LANG']['fideid-form']['bemerkungen'] = array('Bemerkungen', 'Bemerkungen, Hinweise oder Fragen');
+
+// Seite 8
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite8'] = '<h2>Zusammenfassung</h2>';
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite8_summary1'] = 
+'<ul>
+<li>Länderkennung GER gewünscht: <b>#germany#</b></li>
+<li>Datenweitergabe an FIDE gewünscht: <b>#datenschutz#</b></li>
+<li>Antragsteller ist 18 Jahre oder älter: <b>#ueber18#</b></li>
+</ul>';
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite8_summary2'] = '<h3>Beantragte Person</h3>';
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite8_summary2text'] = 
+'<ul>
+<li>Vorname: <b>#vorname#</b></li>
+<li>Nachname: <b>#nachname#</b></li>
+<li>Geschlecht: <b>#geschlecht#</b></li>
+<li>Geburtsdatum: <b>#geburtsdatum#</b></li>
+<li>E-Mail: <b>#email#</b></li>
+</ul>';
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite8_summary3'] = '<h3>Antragsteller</h3>';
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite8_summary3text'] = 
+'<ul>
+<li>Vorname: <b>#vorname#</b></li>
+<li>Nachname: <b>#nachname#</b></li>
+<li>E-Mail: <b>#email#</b></li>
+</ul>';
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite8_summary4'] = '<h3>Weitere Daten zur beantragten Person</h3>';
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite8_summary4text'] = 
+'<ul>
+<li>Turnier/Veranstaltung: <b>#turnier#</b></li>
+<li>Verein: <b>#verein#</b></li>
+<li>Ausweis/Geburtsurkunde: <b>#ausweis#</b></li>
+</ul>';
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite8_summary5'] = '<h3>Bemerkungen</h3>';
+$GLOBALS['TL_LANG']['fideid-form']['antrag_seite8_summary5text'] = 
+'<ul>
+<li><b>#bemerkungen#</b></li>
+</ul>
+<p><b>Sie erhalten nach dem Absenden des Formulars eine E-Mail mit den eingegebenen Daten und einem Link, über den Sie Ihren Antrag freischalten müssen.</b></p>';
+$GLOBALS['TL_LANG']['fideid-form']['submit_all'] = array('Antrag absenden', 'Antrag absenden');