+++ src/Resources/contao/languages/de/modules.php
@@ -13,4 +13,5 @@ $GLOBALS['TL_LANG']['MOD']['fideid'] = array('FIDE-ID-Nummern', 'FIDE-Identifika
 /**
  * Frontend-Module
  */
-$GLOBALS['TL_LANG']['FMD']['fideid'] = array('FIDE-ID Antragsbestätigung', 'FIDE-ID Antragsbestätigung anzeigen');
+$GLOBALS['TL_LANG']['FMD']['fideid_form'] = array('FIDE-ID Antragsformular', 'FIDE-ID Antragsformular anzeigen');
+$GLOBALS['TL_LANG']['FMD']['fideid_confirm'] = array('FIDE-ID Antragsbestätigung', 'FIDE-ID Antragsbestätigung anzeigen');