+++ src/Resources/contao/languages/de/tl_settings.php
@@ -10,3 +10,5 @@ $GLOBALS['TL_LANG']['tl_settings']['fideidverwaltung_legend']       = 'FIDE-ID-N
  */
 $GLOBALS['TL_LANG']['tl_settings']['fideidverwaltung_absender'] = array('Absender für E-Mail-Versand', 'Absender (Name und E-Mail) für den E-Mail-Versand');
 $GLOBALS['TL_LANG']['tl_settings']['fideidverwaltung_mailsignatur'] = array('Signatur für E-Mail-Versand', 'Signatur für den E-Mail-Versand');
+$GLOBALS['TL_LANG']['tl_settings']['fideidverwaltung_bestaetigungsseite'] = array('Bestätigungsseite', 'Seite auf der das Modul für die Antragsbestätigung eingebunden ist.');
+$GLOBALS['TL_LANG']['tl_settings']['fideidverwaltung_uploads'] = array('Ordner für Uploads', 'Ordner für die Uploads von Ausweisen und Geburtsurkunden.');